Create run.bat
create bat file that had a path and drive var. Best practice to use this on the desktop or in another file is to create a shortcut of run.bat as if you use the original run.bat in a different folder from the code it will not work as the path var is for the folder that the .bat is in.